Modernize QGLikelihoodESProducer
- Merge source and header and move to plugins
- Use ESGetToken
